* Added validation to base class * Changed stop inputs to int * Fixed all strategies to use my custom base class * Added new "None" stop loss rule

+++ MetaTraderStrategies/Classes/CBigMove.cpp
@@ -9,9 +9,9 @@ public:
     (
         double          inpLots = 1,
         STOPLOSS_RULE   inpStopLossRule = StaticPipsValue,
-        double          inpStopLossPips = 15,
+        int             inpStopLossPips = 15,
         bool            inpUseTakeProfit = true,
-        double          inpTakeProfitPips = 30,
+        int             inpTakeProfitPips = 30,
         STOPLOSS_RULE   inpTrailingStopLossRule = StaticPipsValue,
         int             inpTrailingStopPips = 20,
         bool            inpGoLong = true,
@@ -60,9 +60,9 @@ CBigMove::~CBigMove(void)
 int CBigMove::Init(
     double          inpLots,
     STOPLOSS_RULE   inpStopLossRule,
-    double          inpStopLossPips,
+    int             inpStopLossPips,
     bool            inpUseTakeProfit,
-    double          inpTakeProfitPips,
+    int             inpTakeProfitPips,
     STOPLOSS_RULE   inpTrailingStopLossRule,
     int             inpTrailingStopPips,
     bool            inpGoLong,
